socket send data return error

+++ src/include/quectel-ec25/socket-client.h
@@ -86,7 +86,7 @@ QTEL_Status_t QTEL_SockClient_Loop(QTEL_SocketClient_t*);
 void          QTEL_SockClient_SetBuffer(QTEL_SocketClient_t*, void *buffer);
 QTEL_Status_t QTEL_SockClient_Open(QTEL_SocketClient_t*, QTEL_HandlerTypeDef*);
 QTEL_Status_t QTEL_SockClient_Close(QTEL_SocketClient_t*);
-uint16_t      QTEL_SockClient_SendData(QTEL_SocketClient_t*, uint8_t *data, uint16_t length);
+int           QTEL_SockClient_SendData(QTEL_SocketClient_t*, uint8_t *data, uint16_t length);
 
 
 #endif /* QTEL_EN_FEATURE_SOCKET */
